// apps/main/src/models/v2/iam/invitationSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { SubjectType } from './subjectType';

/**
 * Summary details of an invitation for this organization.
 */
export interface InvitationSummary {
  /** The invitation id. */
  id: string;
  /** When the invitation was generated and sent. */
  created_at: string;
  /** When the invitation expires and must be sent again. By default, invitations are only valid for 7 days until they must be resent. */
  expires_at: string;
  /** The user id that created the invitation. */
  created_by: string;
  /** The display name of the user who created the invitation. */
  created_by_display_name: string;
  /** The primary email address of the user who created the invitation. */
  created_by_primary_email_address?: string;
  /** The email address this invitation was sent too. */
  email_address: string;
  /** The membership subject type that the invitation is valid for. */
  membership_subject_type: SubjectType;
  /** The membership subject that the invitation is valid for */
  membership_subject: string;
}